전자결재 file copy 시 디렉토리인지 파일인지 여부 파악 추가, log 강화

+++ src/main/java/egovframework/ezEKP/ezApprovalG/web/EzApprovalGController.java
@@ -1,13 +1,20 @@
 package egovframework.ezEKP.ezApprovalG.web;
 
+import java.awt.Graphics;
+import java.awt.GraphicsEnvironment;
+import java.awt.image.BufferedImage;
+import java.awt.image.ImageObserver;
+import java.io.BufferedWriter;
 import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.net.URL;
 import java.net.URLEncoder;
 import java.security.PrivateKey;
 import java.util.ArrayList;
@@ -16,9 +23,12 @@ import java.util.Locale;
 import java.util.Properties;
 
 import javax.annotation.Resource;
+import javax.imageio.ImageIO;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.swing.JEditorPane;
+import javax.swing.JScrollPane;
 
 import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.io.FileUtils;
@@ -58,6 +68,7 @@ import egovframework.let.user.login.vo.LoginVO;
 import egovframework.let.utl.fcc.service.CommonUtil;
 import egovframework.let.utl.fcc.service.EgovDateUtil;
 import egovframework.let.utl.sim.service.EgovFileScrty;
+import gui.ava.html.image.generator.HtmlImageGenerator;
 
 /** 
  * @Description [Controller] 사용자 - 전자결재G
@@ -1611,11 +1622,12 @@ public class EzApprovalGController extends EgovFileMngUtil{
 	@RequestMapping(value = "/ezApprovalG/opinionSave.do", produces = "text/xml;charset=utf-8")
 	@ResponseBody
 	public String opinionSave(@CookieValue("loginCookie") String loginCookie, LoginVO userInfo, @RequestBody String xmlDom) throws Exception{
+		logger.debug("opinionSave Start");
 		userInfo = commonUtil.aprUserInfo(loginCookie);
 		
 		Document docXML = commonUtil.convertStringToDocument(xmlDom);
 		String result = ezApprovalGService.updateOpinionInfo(docXML, userInfo.getCompanyID(), userInfo.getLang(), userInfo.getTenantId());
-		
+		logger.debug("opinionSave End");
 		return result;
 	}
 
@@ -1833,9 +1845,11 @@ public class EzApprovalGController extends EgovFileMngUtil{
 		        }
 			File file = new File(dirPath + userInfo.getCompanyID() + commonUtil.separator + "tempUploadFile" + commonUtil.separator + fileName);
 			
-			if (file.exists()) {
-				FileUtils.copyFile(file, new File(upd + fileName));
-			}  
+			if (file.isFile()) {
+				if (file.exists()) {
+					FileUtils.copyFile(file, new File(upd + fileName));
+				}  
+			}
 			
 			xmlDom.getElementsByTagName("DATA1").item(k).setTextContent(commonUtil.getUploadPath("upload_approvalG.ROOT", userInfo.getTenantId()) + commonUtil.separator + userInfo.getCompanyID() + commonUtil.separator + "uploadFile" + commonUtil.separator + oldYear + commonUtil.separator + ezApprovalGService.getDocDir(fileDocID) + commonUtil.separator + fileName);
 		}
